refactor and fix orders
