+++ web-api/src/main/kotlin/com/flipperdevices/ifrmvp/backend/route/signal/data/model/IncludedFile.kt
@@ -1,0 +1,6 @@
+package com.flipperdevices.ifrmvp.backend.route.signal.data.model
+
+data class IncludedFile(
+    val fileId: Long,
+    val signalCount: Int
+)
